+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## [Unreleased]
+### Changed
+- check-puppet-last-run.rb: Added critical alerting on failures during summary file processing
 
 ## [1.0.0] - 2016-06-21
 ### Changed